split csv_create and import files into 2 , greatly improves run time and reduced clutter
