+++ .gitignore
@@ -1,3 +1,4 @@
+*.csv
 .Rhistory
 .RData
 /unmatched_taxa/*
