+++ image_client/csv_purge_sql/picturae_purge_script.sql
@@ -10,10 +10,12 @@ DELETE FROM casbotany.agent WHERE TimestampCreated  > (now() - interval 20 day);
 
 
 CREATE TEMPORARY TABLE temp_leaf_nodes AS SELECT TaxonID FROM casbotany.taxon WHERE TaxonID
-                    NOT IN (SELECT DISTINCT ParentID FROM casbotany.taxon);
+                       NOT IN (SELECT DISTINCT ParentID FROM casbotany.taxon
+                       WHERE ParentID IS NOT NULL) AND TimestampCreated >= (now() - interval 20 day);
+
 
     -- Step 2: Delete leaf nodes from temporary table
-DELETE FROM casbotany.taxon WHERE TaxonID IN (SELECT TaxonID FROM temp_leaf_nodes) AND TimestampCreated > (now() - interval 20 day);
+DELETE FROM casbotany.taxon WHERE TaxonID IN (SELECT TaxonID FROM temp_leaf_nodes);
 
 
 DROP TEMPORARY TABLE IF EXISTS temp_leaf_nodes;
