--- image_client/csv_purge_sql/upload_time_stamps.csv
@@ -1,3 +1,0 @@
-StartTime,EndTime,UploadCode
-2023-07-28 09:28:13.238155,2023-07-28 09:28:45.936700,476987
-2023-08-14 13:34:23.691441,2023-08-14 13:35:01.087562,888220
