+++ image_client/duplicates-Botany.txt
@@ -1,0 +1,24 @@
+
+Duplicate: cas999999984.jpg
+	 1.3KB: /users/mdelaroca/documents/sandbox_db/specify-sandbox/web-asset-server/image_client/picturae_img/pic_2023-06-28/./cas999999984.jpg
+	 1.3KB: /users/mdelaroca/documents/sandbox_db/specify-sandbox/web-asset-server/image_client/picturae_img/pic_2023-06-28/./cas999999984.jpg
+
+Duplicate: cas999999985.jpg
+	 1.3KB: /users/mdelaroca/documents/sandbox_db/specify-sandbox/web-asset-server/image_client/picturae_img/pic_2023-06-28/./cas999999985.jpg
+	 1.3KB: /users/mdelaroca/documents/sandbox_db/specify-sandbox/web-asset-server/image_client/picturae_img/pic_2023-06-28/./cas999999985.jpg
+
+Duplicate: cas999999986.jpg
+	 1.3KB: /users/mdelaroca/documents/sandbox_db/specify-sandbox/web-asset-server/image_client/picturae_img/pic_2023-06-28/./cas999999986.jpg
+	 1.3KB: /users/mdelaroca/documents/sandbox_db/specify-sandbox/web-asset-server/image_client/picturae_img/pic_2023-06-28/./cas999999986.jpg
+
+Duplicate: cas999999982.jpg
+	 1.3KB: /users/mdelaroca/documents/sandbox_db/specify-sandbox/web-asset-server/image_client/picturae_img/pic_2023-06-28/./cas999999982.jpg
+	 1.3KB: /users/mdelaroca/documents/sandbox_db/specify-sandbox/web-asset-server/image_client/picturae_img/pic_2023-06-28/./cas999999982.jpg
+
+Duplicate: cas999999983.jpg
+	 1.3KB: /users/mdelaroca/documents/sandbox_db/specify-sandbox/web-asset-server/image_client/picturae_img/pic_2023-06-28/./cas999999983.jpg
+	 1.3KB: /users/mdelaroca/documents/sandbox_db/specify-sandbox/web-asset-server/image_client/picturae_img/pic_2023-06-28/./cas999999983.jpg
+
+Duplicate: cas999999981.jpg
+	 1.3KB: /users/mdelaroca/documents/sandbox_db/specify-sandbox/web-asset-server/image_client/picturae_img/pic_2023-06-28/./cas999999981.jpg
+	 1.3KB: /users/mdelaroca/documents/sandbox_db/specify-sandbox/web-asset-server/image_client/picturae_img/pic_2023-06-28/./cas999999981.jpg
