--- image_client/picturae_csv/2023-06-28/picturae_specimen(2023-06-28).csv
@@ -1,7 +1,0 @@
-﻿application_batch,csv_batch,object_type,folder_barcode,specimen_barcode,filed_as_family,barcode_info,path_jpg,collector_number,collector_first_name 1,collector_middle_name 1,collector_last_name 1,collector_first_name 2,collector_middle_name 2,collector_last_name 2,collector_first_name 3,collector_middle_name 3,collector_last_name 3,collector_first_name 4,collector_middle_name 4,collector_last_name 4,collector_first_name 5,collector_middle_name 5,collector_last_name 5,Country,State,County,Locality,Verbatim Date,Start Date,End Date,Notes,Feedback
-1,csv_1234,raw_csv,Cover 999999985_1.jpeg,999999985,TRUE,,picturae_img/2023-6-28/CAS999999985.JPG,999994,Jose,Enrique,De La Roca,,,,,,,,,,,,,United States,California,San Francisco County,"Botanical Gardens SF, on east side","25, April 1994",4/25/1994,4/27/1994,,
-1,csv_1234,raw_csv,Cover 999999981_1.jpg,999999981,TRUE,,picturae_img/2023-6-28/CAS999999981.JPG,999999,Joseph,B.,Candel Sr.,Vladimir,Y.,Chemodanov,,,,,,,,,,United States,California,Marin County,2 mi from western side of Mt.Fake,23' May 1944,5/23/1944,5/23/1944,smudged accession,
-1,csv_1234,raw_csv,Cover 999999982_1.jpg,999999982,TRUE,,picturae_img/2023-6-28/CAS999999982.JPG,999998,Mary,P.,Woolson,,,,,,,,,,,,,United States,California,Marin County,14 mi from eastern side of Mt.Fake,23' May 1944,5/24/1944,5/24/1944,smudged accession,
-1,csv_1234,raw_csv,Cover 999999983_1.jpeg,999999983,TRUE,,picturae_img/2023-6-28/CAS999999983.JPG,999987,Jose,Enrique,De La Roca,,,,,,,,,,,,,United States,California,San Bernardino County,"34.223 N, 114.123 E, 2 miles down path from lake Fable parking lot","25, April 1980",4/25/1980,4/27/1980,,
-1,csv_1234,raw_csv,Cover 999999984_1.jpeg,999999984,TRUE,,picturae_img/2023-6-28/CAS999999984.JPG,999988,Jose,Enrique,De La Roca,,,,,,,,,,,,,United States,California,San Bernardino County,"34.223 N, 114.123 E, 2 miles down path from lake Fable parking lot","25, April 1982",4/25/1982,4/27/1982,,
-1,csv_1234,raw_csv,Cover 999999986_1.jpeg,999999986,TRUE,,picturae_img/2023-6-28/CAS999999986.JPG,999982,Jose,Enrique,De La Roca,,,,,,,,,,,,,United States,California,San Bernardino County,"34.223 N, 114.123 E, 2 miles down path from lake Fable parking lot","25, April 1982",4/25/1982,4/27/1982,,
