--- image_client/taxon_check/unmatched_taxa/manual_taxa_2023-06-28.csv
@@ -1,1 +1,0 @@
-Barcode,FullName,AcceptedName
