--- image_client/taxon_check/unmatched_taxa/unmatch_and_typo2023-08-03.csv
@@ -1,3 +1,0 @@
-fullname,Name_matched,Overall_score,Unmatched_terms,barcodes
-Cliemia almedae,Clidemia almedae,0.97922049354994,,1235
-Impatiens banen,Impatiens,0.5,banen,1236
