--- image_client/taxon_check/unmatched_taxa/unmatch_and_typo_2023-08-07.csv
@@ -1,3 +1,0 @@
-CatalogNumber,fullname,name_matched,overall_score,unmatched_terms
-999999983,Quercus x kinselae,Quercus kinseliae,0.98210117101673,
-999999984,Wrongus wrongica,[No match found],,Wrongus wrongica
