--- image_client/taxon_check/unmatched_taxa/unmatch_and_typo_2023-08-09.csv
@@ -1,2 +1,0 @@
-barcodes,fullname,name_matched,overall_score,unmatched_terms
-1234,Monardella linoides subsp. erecta X australis,Monardella linoides subsp. erecta,0.9,X australis
