--- image_client/taxon_check/unmatched_taxa/unmatch_and_typo_2023-08-10.csv
@@ -1,2 +1,0 @@
-CatalogNumber,fullname,name_matched,overall_score,unmatched_terms
-999999983,Q u e r c u s,[No match found],,Q u e r c u s
