--- image_client/test_barcodes/barcode_db.csv
@@ -1,3 +1,0 @@
-CatalogNumber
-000008708
-000580092
